backend/requirements.txt
# backend/requirements.txt

# Core Flask & SQLAlchemy
Flask==2.3.3
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5

# Authentication & Security
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7

# Utilities
python-dotenv==1.0.0
Flask-CORS==4.0.0

# WebSocket Support
Flask-SocketIO==5.3.6

# AI & OCR
spacy>=3.6.0
openai>=0.28.0
pytesseract>=0.3.10

# File Processing
pandas>=2.0.0
openpyxl>=3.1.0

# AI & Machine Learning
openai>=1.0.0
langchain>=0.1.0
transformers>=4.30.0

# Additional Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Development & Testing
pytest>=7.4.0
pytest-flask>=1.2.0
